[Update] Add MapExt & Move Marker Design Func 2
